requirements.txt
numpy~=1.21.2
torch~=1.9.1
matplotlib~=3.4.3
Pillow~=8.3.2
torchvision~=0.10.1
pandas~=1.3.4
statsmodels~=0.13.2